Fix issues due to the usage of HybridRoleManager
